Refactored weight handling to grams and removed obsolete OrderBloc
This commit transitions the application's weight measurement system from floating-point kilograms to integer-based grams to ensure precision in calculations and storage. It also removes the unused `OrderBloc`, refactors rating logic, and improves UI form handling.

Key changes:
-   **Data Models & Precision:**
    -   **Catch & Offer Models:** Converted weight fields (`initialWeight`, `availableWeight`, `weight`, `previousWeight`) from `double` to `int` (grams).
    -   **Review Model:** Updated `ratingValue` to be stored as an `int`.
    -   **Converters:** Added `formatWeight` helper to display integer gram values as formatted Kg strings.

-   **Business Logic & Repositories:**
    -   **Price Calculation:** Updated `OfferRepository` and form logic to calculate `pricePerKg` using integer arithmetic: `(Total Price * 1000) / Weight(g)`.
    -   **Cleanup:** Deleted the unused `features/fisher/logic/order_bloc` directory (Bloc, Event, State).

-   **UI & Forms:**
    -   **Input Handling:** Refactored `OfferDetails`, `ProductDetails`, `CatchDetails`, and `CounterOfferDialog` to convert user input (Kg) to grams before submission and dynamically auto-calculate pricing fields.
    -   **Notifications Screen:** Implemented a `TabController` with badges for "Offers" and "Messages" counts.
    -   **Ratings:** Updated `AnimatedRatingStars` to handle integer ratings and fixed interaction logic.
    -   **Display Components:** Updated `ForSaleCard`, `OfferCard`, `SoldCard`, and `OrderCard` to utilize the new `formatWeight` utility.

-   **Maintenance:**
    -   **Seeder:** Updated `CatchSeeder` to generate test data with weights in grams.
    -   **Config:** Bumped Dart SDK environment constraint to `^3.10.0`.

+++ lib/core/models/offer.dart
@@ -11,11 +11,11 @@ class Offer extends Equatable {
   final String buyerId;
   final int pricePerKg;
   final int price;
-  final double weight;
+  final int weight;
   final OfferStatus status;
   final String dateCreated;
   final int? previousPrice;
-  final double? previousWeight;
+  final int? previousWeight;
   final int? previousPricePerKg;
 
   // 🆕 ADDED FIELD
@@ -96,10 +96,10 @@ class Offer extends Equatable {
   Offer copyWith({
     OfferStatus? status,
     int? previousPrice,
-    double? previousWeight,
+    int? previousWeight,
     int? previousPricePerKg,
     int? price,
-    double? weight,
+    int? weight,
     int? pricePerKg,
     String? dateCreated,
     bool? hasUpdateForFisher,
@@ -185,11 +185,11 @@ class Offer extends Equatable {
     buyerId: m['buyer_id'] as String,
     pricePerKg: (m['price_per_kg'] as num).toInt(),
     price: (m['price'] as num).toInt(),
-    weight: (m['weight'] as num).toDouble(),
+    weight: (m['weight'] as num).toInt(),
     status: offerStatusFromString(m['status'] as String),
     dateCreated: m['date_created'] as String,
     previousPrice: (m['previous_price'] as num?)?.toInt(),
-    previousWeight: (m['previous_weight'] as num?)?.toDouble(),
+    previousWeight: (m['previous_weight'] as num?)?.toInt(),
     previousPricePerKg: (m['previous_price_per_kg'] as num?)?.toInt(),
     // 🆕 EXTRACTED FIELD
     hasUpdateForBuyer: (m['has_update_buyer'] as int) == 1,
